% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cdat.R
\docType{data}
\name{Cdat}
\alias{Cdat}
\title{Precomputed C matrix}
\format{
A matrix with dimensions (l+1) x (l+1)
}
\usage{
data(Cdat)
}
\description{
Coefficients of the quadratic form used by TEPest
}
\keyword{internal}
