% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TEPexp.R
\name{TEPexp}
\alias{TEPexp}
\title{Exponentiality test based on TEP's}
\usage{
TEPexp(X, nrep = 1000)
}
\arguments{
\item{X}{Vector containing the elements of the sample}

\item{nrep}{Integer, Number of Monte Carlo evaluations to estimate the p-value
(defalts to 1000)}

\item{l}{Integer, Number of Laguerre polynomials evaluated at the sample points -1 (defaults to 10)}
}
\value{
A list with
\itemize{
\item \code{"pval"} — The p-value of the sample estimated by MC
\item \code{"est"} — The value of the test statistic
}
}
\description{
Performs a test of exponentiality consistent
and focused to detect Weibul alternatives
}
\details{
The tests statistics are quadratic forms in the vector of evaluations
of the normalized Laguerre polynomials at the scaled sample points,
introduced by the authors in
\emph{Goodness-of-fit to the Exponential Distribution, focused on Weibull alternatives},
Communications in Statistics, Simulation and Computation \bold{34}, 711-723, 2005.
}
\examples{
\dontrun{
set.seed(42)
X <- rexp(100)
TEPexp(X)
}

}
